Updated to version [0.1.18]

+++ src/app/features/lease-contract/lease-contract-detail/lease-contract-detail.component.ts
@@ -2,6 +2,8 @@ import {Component, OnInit} from '@angular/core';
 import {LeaseContractService} from "../services/lease-contract.service";
 import {LeaseContractBusinessModel} from "../../../common/business-models/lease-contract.business-model";
 import {ActivatedRoute, Router} from "@angular/router";
+import {DialogComponent} from "../../../common/components/dialog/dialog.component";
+import {MatDialog} from "@angular/material/dialog";
 
 @Component({
   selector: 'app-lease-contracts-dashboard',
@@ -17,6 +19,7 @@ export class LeaseContractDetailComponent implements OnInit {
     private router: Router,
     private leaseContractService: LeaseContractService,
     private activatedRoute: ActivatedRoute,
+    private dialog: MatDialog
   ) {
   }
 
@@ -44,17 +47,26 @@ export class LeaseContractDetailComponent implements OnInit {
     this.router.navigate(['/contrats-location/' + this.leaseContractId + '/modification'])
   }
 
-  public delete(): void {
-    this.isLoading = true;
-    this.leaseContractService.deleteLeaseContract(this.leaseContractId).subscribe({
-      next: () => {
-        this.isLoading = false;
-        this.router.navigate(['/biens/' + this.leaseContract?.property.id]);
-      },
-      error: () => {
-        // TODO handle error
-      }
+  public deleteLeaseContract(): void {
+    const message = `Supprimer le contrat conclu entre ${this.leaseContract?.owner?.firstName} ${this.leaseContract?.owner?.lastName}
+    (propriétaire) et ${this.leaseContract?.tenant?.firstName} ${this.leaseContract?.tenant?.lastName} (locataire) engendrera
+    s'il y a, la suppression des paiements effectués par le locataire`
+    const dialogRef = this.dialog.open(DialogComponent, {
+      data: {message},
+      autoFocus: 'dialog'
+    });
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result === "cancel") return;
+      this.isLoading = true;
+      this.leaseContractService.deleteLeaseContract(this.leaseContractId).subscribe({
+        next: () => {
+          this.isLoading = false;
+          this.router.navigate(['/biens/' + this.leaseContract?.property.id]);
+        },
+        error: () => {
+          // TODO handle error
+        }
+      })
     })
   }
-
 }
